requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
gunicorn==25.3.0
psycopg2-binary==2.9.11
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1
reportlab==4.1.0
Pillow==10.3.0
qrcode[pil]==7.4.2
